move graphdriver/projectquota.go to its own package.
